% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{boost_function}
\alias{boost_function}
\title{Internal boost function for Neil's simple boosting/waning model}
\usage{
boost_function(times, pars)
}
\description{
Internal boost function for Neil's simple boosting/waning model
}
