% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{convert_date_to_number}
\alias{convert_date_to_number}
\title{Convert date to integer}
\usage{
convert_date_to_number(x, origin = "2013-01-01")
}
\arguments{
\item{x}{date to be converted}

\item{origin}{character of format "1970-01-01" as date origin}
}
\value{
an integer
}
\description{
Convert date to integer
}
