% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{convert_dist_pars}
\alias{convert_dist_pars}
\title{Distribution mean and variance to par1 and par2 (and vice versa)}
\usage{
convert_dist_pars(
  dist = "gamma",
  par1 = NULL,
  par2 = NULL,
  par3 = NULL,
  mean_par = NULL,
  var_par = NULL
)
}
\arguments{
\item{dist}{character name of the distribution. one of "gamma","beta","beta-binomial","gumbel", "geometric"}

\item{par1}{parameter 1 of the distribution (usually shape or alpha)}

\item{par2}{parameter 2 of the distribution (usually scale or beta)}

\item{par3}{OPTIONAL, but may be the 3rd distribution parameter e.g., N of the beta-binomial}

\item{mean_par}{mean of the distribution}

\item{var_par}{variance of the distribution}
}
\value{
a named vector of the converted parameters
}
\description{
Given parameter 1 and 2 (e.g., mean and variance or shape and scale), finds parameter 1 and parameter 2 (or mean and variance) of the specified distribution. Supported distributions are: gamma, beta, beta-binomial, gumbel, geometric
}
