% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{convert_number_to_date}
\alias{convert_number_to_date}
\title{Convert integer to date}
\usage{
convert_number_to_date(x, origin = "2013-01-01")
}
\arguments{
\item{x}{integer to be converted}

\item{origin}{character of format "1970-01-01" as date origin}
}
\value{
a date
}
\description{
Convert integer to date
}
