% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{convert_years_to_quarters}
\alias{convert_years_to_quarters}
\title{Convert years to text quarters}
\usage{
convert_years_to_quarters(times)
}
\arguments{
\item{times}{a vector of years}
}
\value{
a character vector of modified times
}
\description{
Converts a vector of times, as years with decimals, into quarters. Appends "QX-" to the start of the number, where X is the quarter number.
}
