% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{create_cov_mat}
\alias{create_cov_mat}
\title{Creates a covariance matrix from a vector of standard deviations and a correlation matrix}
\usage{
create_cov_mat(sds, cor_mat)
}
\arguments{
\item{sds}{vector of standard deviations}

\item{cor_mat}{a correlation matrix with nrow and ncol the same as length(sds)}
}
\value{
the covariance matrix
}
\description{
Creates a covariance matrix from a vector of standard deviations and a correlation matrix
}
