% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{create_multi_kinetics_posterior_function}
\alias{create_multi_kinetics_posterior_function}
\title{Posterior function pointer}
\usage{
create_multi_kinetics_posterior_function(
  parTab,
  data,
  PRIOR_FUNC,
  model_func,
  use_log = FALSE,
  ver = "posterior",
  random_effects = TRUE,
  titer_range = NULL,
  solve_likelihood = TRUE
)
}
\arguments{
\item{parTab}{the parameter table controlling information such as bounds, initial values etc.}

\item{data}{the data used for model fitting. Should have variables i (individual ID), time (time since infection/exposure) and titer (measurement).}

\item{PRIOR_FUNC}{pointer to a function specifying priors on model parameters.}

\item{model_func}{pointer to the antibody kinetics model function used}

\item{use_log}{if TRUE, transforms the model predicted titers to log10 titers}

\item{ver}{if "posterior" returns a posterior probability. If "model", returns the model-predicted titers corresponding to the observation times}

\item{random_effects}{if TRUE, assumes that individual-level parameter estimates have random effects}

\item{titer_range}{a vector giving the upper and lower bound of the assay}
}
\description{
Used by lazymcmc::run_MCMC to generate posterior draws from the specified model/data.
}
