% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{dbb}
\alias{dbb}
\title{Beta binomial density}
\usage{
dbb(x, u, v, N)
}
\arguments{
\item{x}{the number of successes}

\item{u}{equivalent to alpha in description}

\item{v}{equivalent to beta in description}

\item{N}{the number of trials}
}
\value{
the probability density of the beta binomial with given input parameters
}
\description{
Beta binomial density
}
\examples{
dbb(10,50,1,1)
}
\seealso{
Other beta_binomial_functions: 
\code{\link{bb_mean}()},
\code{\link{find_bb_2}()},
\code{\link{find_prior_alpha_beta}()},
\code{\link{hist_rbb}()},
\code{\link{pbb}()},
\code{\link{qbb}()},
\code{\link{rbb}()}
}
\concept{beta_binomial_functions}
