% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{extract_numbers_and_text}
\alias{extract_numbers_and_text}
\title{Separates text from numbers in square brackets}
\usage{
extract_numbers_and_text(
  dat,
  target_var,
  number_name = "number",
  text_name = "text"
)
}
\arguments{
\item{dat}{a tibble or data frame}

\item{target_var}{the variable name as a string to be split}

\item{number_name}{the name to give to the new number variable}

\item{text_name}{the name to give to the new text variable}
}
\value{
a modified tibble
}
\description{
Separates text from numbers in square brackets
}
