% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{extract_population_distributions}
\alias{extract_population_distributions}
\title{Extracts posterior draws for the population-level parameters}
\usage{
extract_population_distributions(chain, parTab, nsamp)
}
\arguments{
\item{chain}{the MCMC chain object to draw posterior samples from}
}
\value{
a tibble with posterior draws
}
\description{
Extracts posterior draws for the population-level parameters
}
