% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fit_distribution_quick}
\alias{fit_distribution_quick}
\title{Fit distribution to data vector}
\usage{
fit_distribution_quick(
  dist = "normal",
  x,
  ver = "draws",
  lower_bound = NULL,
  upper_bound = NULL,
  control = list(abstol = 1e-08, reltol = 1e-08)
)
}
\arguments{
\item{dist}{character name of the distribution}

\item{x}{vector of data to be fitted to}

\item{ver}{either "draws" or "density". If "draws", estimates the parameters directly. If "density", fits the distribution to the density of x.}

\item{lower_bound}{OPTIONAL lower bound of optimization (uses Brent method)}

\item{upper_bound}{OPTIONAL upper bound of optimization (uses Brent method)}

\item{control}{list of controls for optim}
}
\value{
a vector of the estimated parameters and a ggplot2 object showing the fit
}
\description{
Given parameter 1 and 2 (e.g., mean and variance or shape and scale), finds parameter 1 and parameter 2 (or mean and variance) of the specified distribution. A key use case is to fit to posterior draws to create priors for downstream MCMC. This uses optim (default Nelder-Mead). Supported distributions are: normal, gamma, beta, log-normal, geometric, poisson
}
