% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fit_geometric}
\alias{fit_geometric}
\title{Function for optim to fit a geometric distribution}
\usage{
fit_geometric(prob, dat)
}
\arguments{
\item{prob}{probability of success}

\item{dat}{is a vector of event observation times}
}
\value{
negative sum log likelihood from dgeom
}
\description{
Function for optim to fit a geometric distribution
}
