% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{from_unit_scale}
\alias{from_unit_scale}
\title{from_unit_scale}
\usage{
from_unit_scale(x, min = 1, max = 100, logflag = FALSE, logbase = 10)
}
\arguments{
\item{x}{numeric value to be transformed}

\item{min}{the bottom of the linear scale. Defaults to 1}

\item{max}{the top of the linear scale. Defaults to 100}

\item{logflag}{Optional flag to convert to a unit scale via the log scale. Defaults to FALSE}

\item{logbase}{Optional numeric value to be used as the base for the log scale}
}
\value{
numeric value with the converted input
}
\description{
Transforms a given value from the scale of 0-1 to min-max, either directly to a linear scale or via a log scale
}
\examples{
from_unit_scale(0.5,20,80)
from_unit_scale(0.33,0,16,TRUE,10)
}
\seealso{
\code{\link{toUnitScale}} inverse function
}
