% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_function.R
\name{generalized_logistic_function}
\alias{generalized_logistic_function}
\title{Generalized logistic function}
\usage{
generalized_logistic_function(
  x,
  A = 0,
  K = 1,
  C = 1,
  Q = 1,
  k = 0,
  M = 0,
  v = 1
)
}
\arguments{
\item{x}{the vector of values to solve the function over}

\item{A}{the lower asymptote}

\item{K}{the upper asymptote, also the carrying capacity when A=0 and C=1}

\item{C}{Usually C=1, but if C != 1 then the upper asymptote = A + (K-A)/(C^(1/v))}

\item{Q}{related to the initial value at Y(0)}

\item{k}{IMPORTANT: the growth rate}

\item{M}{the starting time ie. shifts in the x-axis}

\item{v}{v>0, affecting where maximum growth occurs}
}
\value{
a vector of values for each entry of x, giving y(x) where y() is the generalized logistic function.
}
\description{
Implementation of the fully generalized logistic function, as described at \url{https://en.wikipedia.org/wiki/Generalised_logistic_function}.
Note that the default arguments are set to give the logistic function, but each parameter can be customized.
}
\examples{
y <- generalized_logistic_function(x=seq(-10,10,by=0.1), k=0.5)
}
