% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{get_beta_mode}
\alias{get_beta_mode}
\title{Find beta distribution mode}
\usage{
get_beta_mode(alpha, beta)
}
\arguments{
\item{alpha}{alpha parameter}

\item{beta}{beta parameter}
}
\value{
the mode of the distribution
}
\description{
Find beta distribution mode
}
