% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{get_days_per_month}
\alias{get_days_per_month}
\title{Get days per month of the year}
\usage{
get_days_per_month(breaks = 12)
}
\arguments{
\item{breaks}{total number of months to consider (defaults to 12 to give raw days per each month)}
}
\value{
vector of days
}
\description{
Returns a vector of days per month
}
