% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{get_kinetics_prediction_intervals}
\alias{get_kinetics_prediction_intervals}
\title{Calculate posterior prediction intervals for model trajectories}
\usage{
get_kinetics_prediction_intervals(
  model_func,
  n_samp,
  indivs = NULL,
  chain,
  dat,
  expand_times = TRUE,
  times = NULL,
  use_log = TRUE,
  add_noise = FALSE,
  titer_range = NULL
)
}
\arguments{
\item{model_func}{pointer to antibody kinetics function used to general modeled trajectories}

\item{n_samp}{how many posterior draws to use?}

\item{indivs}{optional, if not NULL, a vector of which individual indices to calculate trajectories for}

\item{chain}{the MCMC chain object to draw posterior samples from}

\item{dat}{the data used in the original model fitting}

\item{expand_times}{if TRUE, solves the model over all times specified in \code{times}, otherwise just solves the model for the observation times}

\item{times}{vector of times to solve model over if expand_times is TRUE}

\item{use_log}{if TRUE, log10 transforms the model-predicted titers}

\item{add_noise}{if TRUE, adds simulated noise from the observation error distribution to the trajectories (and thus can be used to get prediction rather than credible intervals)}

\item{titer_range}{a vector giving the upper and lower bounds of the assay}
}
\value{
a list with two tibbles. The first contains n_samp posterior draws for the model trajectories. The second returns posterior median and 95\% CrI or prediction intervals for the model trajectories.
}
\description{
Calculate posterior prediction intervals for model trajectories
}
