% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{kinetics_ode}
\alias{kinetics_ode}
\title{Boosting and waning model based on within-host ODEs}
\usage{
kinetics_ode(times, pars)
}
\arguments{
\item{times}{vector of times to solve model over}

\item{pars}{vector of named model parameters}
}
\value{
vector of antibody levels
}
\description{
Implements the analytical solution to the ODE model described in Pelleau et al. 2021 (JID).
}
\details{
Model parameters:
\itemize{
\item{"A"}{initial antibody level}
\item{"beta"}{growth rate of antibody secreting cells}
\item{"rho"}{proportion of B cells which are short-lived}
\item{"delta"}{time at which antibody levels start increasing}
\item{"c_l"}{decay rate of long-lived B cells}
\item{"c_s"}{decay rate of short-lived B cells}
\item{"r"}{antibody decay rate}
}
}
\examples{
pars <- c("A"=0,"beta"=100,"rho"=0.25,"c_s"=0.15,"c_l"=0.0075,"delta"=0,"r"=0.001)
times <- seq(0,1000,by=0.1)
y <- kinetics_ode(times,pars)
plot(times, y, type='l')
}
\seealso{
Other kinetics_models: 
\code{\link{kinetics_gamma}()},
\code{\link{kinetics_piecewise}()},
\code{\link{kinetics_power_function}()},
\code{\link{kinetics_ranjeva}()},
\code{\link{kinetics_simple}()}
}
\concept{kinetics_models}
