% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{kinetics_piecewise}
\alias{kinetics_piecewise}
\title{Piecewise linear model of boosting and biphasic waning}
\usage{
kinetics_piecewise(times, pars)
}
\arguments{
\item{times}{vector of times to solve model over}

\item{pars}{vector of named model parameters}
}
\value{
vector of antibody levels
}
\description{
Simple piecewise linear model of linear antibody boosting on log scale followed by linear biphasic waning. Model as in Hay et al. 2019 (PLOS Comp Biol).
}
\details{
Model parameters:
\itemize{
\item{"t_i"}{Timing of infection}
\item{"tp"}{Time of peak antibody level post infection}
\item{"mu"}{Magnitude of boost}
\item{"dp"}{Proportion of boost lost in short-term waning}
\item{"ts"}{Duration of short term waning phase}
\item{"m"}{Long-term antibody waning rate}
\item{"lower_bound"}{Lower bound of allowable antibody levels}
\item{"y0"}{True starting titer level}
}
}
\examples{
pars <- c("tp"=21,"mu"=8,"dp"=0.2,"ts"=50,"m"=0.0075,"lower_bound"=0, "y0"=0)
times <- seq(0,100,by=0.1)
y <- kinetics_piecewise(times,pars)
plot(times, y, type='l')
}
\seealso{
Other kinetics_models: 
\code{\link{kinetics_gamma}()},
\code{\link{kinetics_ode}()},
\code{\link{kinetics_power_function}()},
\code{\link{kinetics_ranjeva}()},
\code{\link{kinetics_simple}()}
}
\concept{kinetics_models}
