% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{kinetics_power_function}
\alias{kinetics_power_function}
\title{Power function boosting and waning model}
\usage{
kinetics_power_function(times, pars)
}
\arguments{
\item{times}{vector of times to solve model over}

\item{pars}{vector of named model parameters}
}
\value{
vector of solved antibody levels
}
\description{
Solve the model described in Teunis et al. 2016 (Epidemics) and used in Aiemjoy et al. 2022 (Lancet Microbe). There is an addition of a titer-dependent boosting term as used in Ranjeva et al. 2019 (Nature Communications).
}
\details{
Model parameters:
\itemize{
\item{"t_peak"}{time of peak antibody level post exposure}
\item{"peak"}{peak antibody level on natural scale}
\item{"k"}{antibody level dependent boosting term}
\item{"y0"}{starting antibody level on natural scale -- Note this needs to start at 1}
\item{"v"}{antibody decay rate}
\item{"r"}{shape of antibody decay function}
}
}
\examples{
N <- 25
r <- rlnorm(N, log(1.3),0.01) ## Shape parameter
v <- rexp(N,500) ## Decay rate
y0 <- 1## Start titre
k <- 0 ## Titre dependency
peak <- 10^(rnorm(N,5, 1.5)) ## Peak titre

times <- seq(0,365,by=0.1)
ys <- matrix(nrow=N,ncol=length(times))
for(i in 1:N){
  ys[i,] <- kinetics_power_function(times, c(t_peak=21,peak=peak[i],r=r[i],v=v[i],y0=1,k=0))
}
ys <- log10(ys)

plot(times,(ys[1,]),type='l',ylim=c(0,10))
for(i in 2:N){
  lines(times,(ys[i,]))
}
}
\seealso{
Other kinetics_models: 
\code{\link{kinetics_gamma}()},
\code{\link{kinetics_ode}()},
\code{\link{kinetics_piecewise}()},
\code{\link{kinetics_ranjeva}()},
\code{\link{kinetics_simple}()}
}
\concept{kinetics_models}
