% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kinetics_power_function_cpp}
\alias{kinetics_power_function_cpp}
\title{Fast implementation of the power-law waning antibody kinetics function}
\usage{
kinetics_power_function_cpp(times, pars)
}
\arguments{
\item{times}{vector of times to solve model over}

\item{pars}{vector of named model parameters}
}
\description{
Fast implementation of the power-law waning antibody kinetics function
}
