% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{kinetics_ranjeva}
\alias{kinetics_ranjeva}
\title{Boosting and waning model with ceiling effect and asymptoting boost}
\usage{
kinetics_ranjeva(times, pars, boost = f_rise, wane = f_wane)
}
\arguments{
\item{times}{vector of times to solve model over}

\item{pars}{vector of named model parameters}

\item{boost}{pointer to the antibody boosting function (defaults to f_rise)}

\item{wane}{pointer to the antibody waning function (defaults to f_wane)
#' @details Model parameters:
\itemize{
\item{"t_peak"}{time of peak antibody level post exposure}
\item{"peak"}{peak antibody level on natural scale}
\item{"k"}{antibody level dependent boosting term}
\item{"y0"}{starting antibody level on natural scale}
\item{"w"}{antibody decay rate}
\item{"frac"}{proportion of the intial boost lost during short-term waning}
}}
}
\value{
vector of antibody levels
}
\description{
Implements the antibody kinetics model from Ranjeva et al. 2019 (Nature Communications). Key features of this model are a short-term, titer-dependent boost that asymptotes towards a peak value rather than continuing to grow exponentially. The waning phase is biphasic -- an initial proportion of the boost is lost and the asymptotes to a steady state.
}
\examples{
pars <- c(y0=1,peak=80,k=0,t_peak=28,w=0.008,frac=0.5)
times <- seq(0,1000,by=0.1)
y <- kinetics_ranjeva(times,pars)
plot(times, y, type='l')
}
\seealso{
Other kinetics_models: 
\code{\link{kinetics_gamma}()},
\code{\link{kinetics_ode}()},
\code{\link{kinetics_piecewise}()},
\code{\link{kinetics_power_function}()},
\code{\link{kinetics_simple}()}
}
\concept{kinetics_models}
