% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{kinetics_simple}
\alias{kinetics_simple}
\title{Simple boosting and waning kinetics}
\usage{
kinetics_simple(
  times,
  pars,
  boost = boost_function,
  wane = wane_function_asymptote
)
}
\arguments{
\item{times}{vector of times to solve model over}

\item{pars}{vector of named model parameters}

\item{boost}{pointer to the function used to calculate boosting (defaults to boost_function)}

\item{wane}{pointer to the function used to calculate waning (defaults to wane_function_asymptote)}
}
\value{
vector of antibody levels
}
\description{
Implements the biphasic boosting and waning model from Singanayagam et al. 2021 (Lancet ID). A phenomonelogical assuming exponential growth and decline, where the weighting of the two components shifts from growth to decline, swapping over at the peak time. Can have an asymptote or not in the waning rate. NOTE: when using the asymptote waning function, the function \emph{does not} peak at t_peak, but slightly after.
}
\details{
Model parameters:
\itemize{
\item{"t_peak"}{time of peak antibody level post exposure}
\item{"ymax"}{peak antibody level on natural scale}
\item{"a"}{boosting rate}
\item{"b"}{waning rate}
\item{"c"}{asymptote of waning (can set to 0, or use wane_function instead of wane_function_asymptote to decline to 0)}
}
}
\examples{
pars<- c(a=0.3,b=0.05,c=0.75,ymax=100,t_peak=18)
times <- seq(0,100,by=1)
y <- kinetics_simple(times,pars)
plot(times, y, type='l')
}
\seealso{
Other kinetics_models: 
\code{\link{kinetics_gamma}()},
\code{\link{kinetics_ode}()},
\code{\link{kinetics_piecewise}()},
\code{\link{kinetics_power_function}()},
\code{\link{kinetics_ranjeva}()}
}
\concept{kinetics_models}
