% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(p)
}
\arguments{
\item{p}{numeric value to be converted}
}
\value{
numeric value of the converted input
}
\description{
Converts a given value from the scale of -Inf and +Inf to 0 and 1
}
\examples{
logistic(56)
logistic(logit(0.5))
}
\seealso{
\code{\link{logistic}} returns the inverse value (ie. the logit function)
}
