% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{logistic_function_full}
\alias{logistic_function_full}
\title{Logistic function with start and end}
\usage{
logistic_function_full(t, start, end, rate, switch = 100)
}
\arguments{
\item{t}{value to solve over}

\item{start}{the start value}

\item{end}{the end value}

\item{rate}{rate of exponential growth}

\item{switch}{the midpoint of the logistic function}
}
\value{
numeric value
}
\description{
The logistic function with specified start, end, rate and switchpoint
}
\examples{
logistic_function_full(seq(0,100,by=1),0.3,0.9,0.1,50)
}
