% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{numeric value to be converted}
}
\value{
numeric value of the converted input
}
\description{
Converts a given value from the scale of 0 and 1 to between -Inf and Inf
}
\examples{
logit(0.5)
}
\seealso{
\code{\link{logistic}} returns the inverse value (ie. the logistic function)
}
