% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{median_quantile}
\alias{median_quantile}
\title{Median and quantiles}
\usage{
median_quantile(x)
}
\arguments{
\item{x}{the vector}
}
\value{
vector with 95\% quantiles and median
}
\description{
Calculates the median and 95\% quantiles from a vector
}
