% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{plot_model_fits}
\alias{plot_model_fits}
\title{Plots posterior estimates for the latent antibody trajectories}
\usage{
plot_model_fits(
  model_func,
  n_samp,
  indivs = NULL,
  chain,
  dat,
  expand_times = TRUE,
  times = NULL,
  use_log = TRUE,
  plot_draws = TRUE
)
}
\arguments{
\item{model_func}{pointer to antibody kinetics function used to general modeled trajectories}

\item{n_samp}{how many posterior draws to use?}

\item{indivs}{optional, if not NULL, a vector of which individual indices to calculate trajectories for}

\item{chain}{the MCMC chain object to draw posterior samples from}

\item{dat}{the data used in the original model fitting}

\item{expand_times}{if TRUE, solves the model over all times specified in \code{times}, otherwise just solves the model for the observation times}

\item{times}{vector of times to solve model over if expand_times is TRUE}

\item{use_log}{if TRUE, log10 transforms the model-predicted titers}

\item{plot_draws}{if TRUE, then plots individual draws from the posterior, otherwise plots a ribbon of credible/prediction intervals and posterior median.}
}
\value{
a ggplot2 object
}
\description{
Plots posterior estimates for the latent antibody trajectories
}
