% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plots posterior densities and traceplots for the model parameters}
\usage{
plot_posteriors(
  chain,
  parTab,
  nsamp = 100,
  ver = "means",
  indivs = NULL,
  transform = FALSE
)
}
\arguments{
\item{chain}{the MCMC chain object to draw posterior samples from}

\item{ver}{either "means","var",or "indivs" to plot population mean parameter estimates, population variance of parameters, or individual-level parameters.}

\item{indivs}{optional, if not NULL, a vector of which individual indices to calculate trajectories for}

\item{transform}{if TRUE, exponentiates the parameter values.}
}
\value{
a ggplot2 object
}
\description{
Plots posterior densities and traceplots for the model parameters
}
