% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{print_estimate}
\alias{print_estimate}
\title{Prints the average and quantiles of a vector}
\usage{
print_estimate(x, average = "mean", quantiles = c(0.025, 0.975), nsignif = 3)
}
\arguments{
\item{x}{the vector of values to summarize}

\item{average}{character describing the average version to use (mean or median)}

\item{quantiles}{vector of length 2 giving the quantiles to calculate}

\item{nsignif}{number of significant figures to report}
}
\value{
a string with the formatted estimate
}
\description{
Prints the average and quantiles of a vector
}
