% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{rbb}
\alias{rbb}
\title{Beta binomial random generation}
\usage{
rbb(n, u, v, N)
}
\arguments{
\item{n}{number of draws}

\item{u}{equivalent to alpha in description}

\item{v}{equivalent to beta in description}

\item{N}{the number of trials}
}
\value{
the beta binomial random draw function
}
\description{
Beta binomial random generation
}
\examples{
rbb(100,50,1,1)
}
\seealso{
Other beta_binomial_functions: 
\code{\link{bb_mean}()},
\code{\link{dbb}()},
\code{\link{find_bb_2}()},
\code{\link{find_prior_alpha_beta}()},
\code{\link{hist_rbb}()},
\code{\link{pbb}()},
\code{\link{qbb}()}
}
\concept{beta_binomial_functions}
