% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plots}
\alias{save_plots}
\title{Saves a pdf and png of a ggplot2 object}
\usage{
save_plots(p, wd, save_name, width, height)
}
\arguments{
\item{p}{the ggplot2 object to be saved}

\item{wd}{the full working directory to save to}

\item{save_name}{the filename, without file extension}

\item{width}{the width of the plot in inches}

\item{height}{the height of the plot in inches}
}
\description{
Saves a pdf and png of a ggplot2 object
}
