% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_model.R
\name{sir_model}
\alias{sir_model}
\title{Simple SIR model implementation}
\usage{
sir_model(
  pars = c(R0 = 1.5, gamma = 1/5),
  times = seq(0, 365, by = 0.1),
  initial_states = c(S = 1 - 1e-05, I = 1e-05, R = 0)
)
}
\arguments{
\item{pars}{named vector of model parameters, give R0 and gamma}

\item{times}{vector of times to solve model over}

\item{initial_states}{named vector of initial states, giving values for S, I and R. If the sum is more than 1, uses frequency dependent model}
}
\value{
a data frame with the solved SIR model, including compartment sizes over time, incidence and growth rate of incidence
}
\description{
Simple SIR model implementation
}
