% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_no_x_axis}
\alias{theme_no_x_axis}
\title{ggplot2 theme no x axis}
\usage{
theme_no_x_axis()
}
\value{
A ggplot2 theme object
}
\description{
Removes the x-axis line, ticks and text from a ggplot2 object
}
\keyword{ggplot2,}
\keyword{theme}
