% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_overall}
\alias{theme_overall}
\title{My overall ggplot2 theme}
\usage{
theme_overall()
}
\value{
A ggplot2 theme object
}
\description{
Adds my custom ggplot2 theme to a ggplot2 object, with the intention of standardizing my plots.
}
\examples{
ggplot(data=data.frame(x=rnorm(100))) + geom_histogram(aes(x=x)) + theme_overall()
}
\keyword{ggplot2,}
\keyword{theme}
