% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{to_unit_scale}
\alias{to_unit_scale}
\title{to_unit_scale}
\usage{
to_unit_scale(y, min = 1, max = 100, logflag = FALSE, logbase = 10)
}
\arguments{
\item{y}{numeric value to be transformed}

\item{min}{the bottom of the linear scale. Defaults to 1}

\item{max}{the top of the linear scale. Defaults to 100}

\item{logflag}{Optional flag to convert to a unit scale via the log scale. Defaults to FALSE}

\item{logbase}{Optional numeric value to be used as the base for the log scale}
}
\value{
numeric value with the converted input
}
\description{
Transforms a given value from the scale of min-max to 0-1, either directly from a linear scale or via a log scale
}
\examples{
to_unit_scale(65,20,80)
to_unit_scale(15,0,16,TRUE,10)
}
\seealso{
\code{\link{fromUnitScale}} inverse function
}
