% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{transform_logistic}
\alias{transform_logistic}
\title{Custom logistic transformation}
\usage{
transform_logistic(x, xmin, xmax)
}
\arguments{
\item{x}{numeric value to be scaled}

\item{xmin}{numeric value representing the lower bound of the new scale}

\item{xmax}{numeric value representing the upper bound of the new scale}
}
\value{
numeric value with the converted input
}
\description{
Transforms a given value from between -Inf and Inf to xmin and xmax
}
\examples{
transform_logistic(5,3,10)
}
\seealso{
\code{\link{transform_logit}} inverse function
}
