% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{transform_logit}
\alias{transform_logit}
\title{Custom logit transform}
\usage{
transform_logit(x, xmin, xmax)
}
\arguments{
\item{x}{numeric value to be scaled}

\item{xmin}{numeric value representing the lower bound of the old scale}

\item{xmax}{numeric value representing the upper bound of the old scale}
}
\value{
numeric value with the converted input
}
\description{
Transforms a given value from between to xmin and xmax to -Inf and Inf
}
\examples{
transform_logit(0,3,10)
}
\seealso{
\code{\link{transform_logistic}} inverse function
}
