% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{wane_function}
\alias{wane_function}
\title{Internal wane function for Neil's simple boosting/waning model}
\usage{
wane_function(times, pars)
}
\description{
Internal wane function for Neil's simple boosting/waning model
}
