% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_kinetics.R
\name{wane_function_asymptote}
\alias{wane_function_asymptote}
\title{Internal wane function for Neil's simple boosting/waning model, with added asymptote}
\usage{
wane_function_asymptote(times, pars)
}
\description{
Internal wane function for Neil's simple boosting/waning model, with added asymptote
}
