% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_nice_axes}
\alias{theme_nice_axes}
\title{ggplot2 theme nice axes}
\usage{
theme_nice_axes()
}
\value{
A ggplot2 theme object
}
\description{
Makes some nice axis lines for a ggplot2 object
}
\examples{
ggplot(data=data.frame(x=rnorm(100))) + geom_histogram(aes(x=x)) + theme_nice_axes()
}
\keyword{ggplot2,}
\keyword{theme}
